Add files via upload
AICP Internship Task Week 4
Hexagon is a geometrical shape having six sides while square is a shape which has four equal sides. You
are required to write Python or C++ program for hexagon and square having data members and
member functions, and create a repeating menu keeping in mind the following requirements: Consider last digit of your CNIC as the length of one side of hexagon. For example, if your CNIC is
XY210351532, the last digit will be 2.
Calculate area and perimeter of hexagon. (Hint: Area of hexagon= 1.5*1.732*s ; where ‘s’ is the
length of one side of hexagon. Perimeter of hexagon= 6*s; where ‘s’ is the length of one side of
hexagon)
Calculate sum of all the angles of hexagon. (Hint: Sum of all the angles of hexagon= 6*a; where ‘a’ is
the measurement of one angle of hexagon which is equal to 120.)
When the input of user is ‘1’, display area, perimeter and sum of all the angles of hexagon.
Calculate area and perimeter of square whereas;
Length of one side of square=last digit of CNIC + 1. (Hint: Area of square= (length)2; where ‘length’
is the length of one side of square. Perimeter of square= 4*length; where ‘length’ is the length of
one side of the square.)
When the input of user is ‘2’, display the area and perimeter of the square.
On any other input, the program should exit.